Extract `ScrollMessage` to its own file
